backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
celery==5.4.0
redis==5.0.8
stripe==8.1.0
pypdf==6.0.0
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
ruff==0.9.9
black==25.1.0
